+++ src/components/TodoList.tsx
@@ -1,3 +1,5 @@
+import { useState, createContext } from 'react';
+
 import { DateCard } from '../organisms';
 
 import { AddTodoDto, ETIndexed } from '../DB/indexed';
@@ -11,7 +13,6 @@ import {
 } from 'react-beautiful-dnd';
 import { useQuery } from '@tanstack/react-query';
 import styled from '@emotion/styled';
-import { useState, createContext } from 'react';
 
 type contextType = [boolean, React.Dispatch<React.SetStateAction<boolean>>];
